+++ app/src/main/res/layout/activity_training_statistics.xml
@@ -144,13 +144,12 @@
                 android:alpha="0.2"
                 android:background="@android:color/black" />
 
-            <com.github.mikephil.charting.charts.BarChart
-                android:layout_marginVertical="10dp"
-                android:id="@+id/speed_bar_chart"
+            <FrameLayout
+                android:id="@+id/speed_line_chart_container"
                 android:layout_width="match_parent"
-                android:layout_height="250dp"
-                android:layout_weight="1"
-                android:visibility="visible" />
+                android:layout_height="wrap_content"
+                android:layout_marginVertical="10dp" />
+
 
             <View
                 android:layout_marginHorizontal="20dp"
