+++ app/src/main/res/layout/fragment_speed_statistics.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <com.github.mikephil.charting.charts.BarChart
+        android:id="@+id/speed_bar_chart"
+        android:layout_width="match_parent"
+        android:layout_height="250dp" />
+
+</FrameLayout>
